requirements.txt
flask
streamlit
openai
tesseract
pytesseract
pillow
matplotlib
pandas
easyocr
PyPDF2
python-dotenv